rename error
https://coveord.atlassian.net/browse/CDX-646
