+++ packages/cli/.coveo/config.json
@@ -1,0 +1,1 @@
+{"version": 1, "organization": ""}
